feat: a few new endpoints and fixes
